+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - [changed] Improved error handling in FCM by mapping more server-side
   errors to client-side error codes.
 
+- [added] The Firebase dynamic links module with the `get_link_stats` API
+
 # v2.9.0